refactor: extract every controller from each router
