+++ packages/api/src/controllers/store.controller.ts
@@ -1,0 +1,13 @@
+import { type Ctx } from '../common';
+
+export const getStoresHandler = async (ctx: Ctx) => {
+  return ctx.prisma.store.findMany();
+};
+
+export const getStoresWithItemsHandler = async (ctx: Ctx) => {
+  return ctx.prisma.store.findMany({
+    include: {
+      items: true,
+    },
+  });
+};
