+++ packages/api/src/router/store.ts
@@ -1,15 +1,8 @@
+import { getStoresHandler, getStoresWithItemsHandler } from '../controllers/store.controller';
 import { createTRPCRouter, publicProcedure } from '../trpc';
 
 export const storeRouter = createTRPCRouter({
-  getAll: publicProcedure.query(({ ctx }) => {
-    return ctx.prisma.store.findMany();
-  }),
+  getAll: publicProcedure.query(({ ctx }) => getStoresHandler(ctx)),
 
-  getAllWithItems: publicProcedure.query(({ ctx }) => {
-    return ctx.prisma.store.findMany({
-      include: {
-        items: true,
-      },
-    });
-  }),
+  getAllWithItems: publicProcedure.query(({ ctx }) => getStoresWithItemsHandler(ctx)),
 });
