+++ packages/api/src/schema/bot.schema.ts
@@ -1,0 +1,13 @@
+import { TypeOf, z } from 'zod';
+
+/*------------------------------------*/
+
+export const sendDiscordMsgInput = z.object({
+  username: z.string(),
+  prUrl: z.string(),
+  coins: z.string(),
+  webhookDiscordUrl: z.string(),
+});
+export type SendDiscordMsgInputType = TypeOf<typeof sendDiscordMsgInput>;
+
+/*------------------------------------*/
